Extract AR::Core#inspect and AR::Core.inspect's test code from base_test.rb
The methods got moved to core.rb in commit
b2c9ce341a1c907041f55461aefebb0321280cb5, but the tests never did.
